+++ packages/next/server/node-polyfill-readable-stream.js
@@ -1,4 +1,4 @@
-import { ReadableStream } from './web/sandbox/polyfills'
+import { ReadableStream } from './web/sandbox/readable-stream'
 
 // Polyfill ReadableStream in the Node.js environment
 if (!global.ReadableStream) {
